pyproject.toml
[project]
name = "cosmo-rag"
version = "0.1.0"
description = "RAG based system on arxiv papers related to cosmology"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Sreerag", email = "sreeragradhakrishnan01@gmail.com"}
]

keywords = ["cosmology", "arxiv", "rag"]
dependencies = [
    "pymupdf",
    "langchain",
    "langchain-text-splitters",
    "sentence-transformers",
    "faiss-cpu",
    "numpy",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"